requirements.txt
groq==0.20.0
pandas==2.2.3
python-dotenv==1.1.0
